Prevent sprintf from overlapping
Since the src and dst overlapping when calling
`sprintf` is undefined behavior.

I prevent the problem by moving the value of
src pointer. And use `len` to store the length
of the string in the progress of concating,
which can also save time since we don't have to
call `strlen` to calculate it again.