Use computed goto to replace switch-cases
Use label as value featured in GNU extension
to implement computed goto and replace the
condition switch in `src/http_parser.c`.

Use perf and htstress provided in sysprog21/khttpd
to compare the performance of these two version

Both of the cases use htstress with the following
parameters:
    ./htstress -n 100000 -c 1 -t 4 \
	http://localhost:8081/

* for computed goto version:
    Available samples
    12K branch-misses:u
    14K branch-instructions:u

    requests:      100000
    good requests: 100000 [100%]
    bad requests:  0 [0%]
    socker errors: 0 [0%]
    seconds:       3.839
    requests/sec:  26047.527

* for switch-case version:
    Available samples
    13K branch-misses:u
    14K branch-instructions:u

    requests:      100000
    good requests: 100000 [100%]
    bad requests:  0 [0%]
    socker errors: 0 [0%]
    seconds:       4.000
    requests/sec:  24997.963

from the two cases above we can see that
the number of branch misses and the execution
time is reduced.

Note: storing address of labels in an array
may cause cppcheck to report "Syntax Error"