Fix to send the right content size
To send the right information about the content,
we need to specify the content size by using the
filesize variable we get from struct stat.

But to write an null character at the end of the
packet in the meanwhile, we should send
filesize + 1 bytes to include the null character.