Fix slowdown error when accessing non-exist file
When the client trying to access some files that do not
exist in the webroot directory, it may cause the server
to suffer performance slowdown.

Instead of waiting the client to send another data in,
if we find out that the file does not exist, we send the
error message about non-exist file and close the connection.