Revise the compare of double with bitops
To compare a double is greater than the other one, we
must first compare the exponent part, if exponent parts
are same, then we compare their fraction parts.

Since we are comparing two positive double and
the exponent part is placed in more significant bits
than fraction part, we can compare them by differencing
them(like a - b) and checking the LSB.

If LSB is 1, it means b is greater than a, otherwise a
is greater than b.